<feat> movement refactor, the 'move*' functions now return the `Action2` values we need to move by
keeping this in the game loop should make it easier to handle collisions and resets with the obstacles and the sides of the canvas

+++ src/2d/my_types.py
@@ -4,7 +4,7 @@ import pygame
 type State = tuple[int, int, int, int]
 
 ## ([-1, 1], [-1, 1]) for delta_x and delta_y
-type Action2 = tuple[int, int]
+type Action2 = tuple[float, float]
 ## Up, Down, Left, Right
 type Action4 = tuple[bool, bool, bool, bool]
